move LayerFixtures to a separate helper class
